refactor: split manager specs and catalog endpoints into dedicated routers
