+++ routers/manager_catalog.py
@@ -1,0 +1,111 @@
+from typing import Optional
+
+from fastapi import APIRouter, Depends, HTTPException, Query
+from sqlalchemy.ext.asyncio import AsyncSession
+
+from core.database import get_session
+from core.security import get_current_username
+from schemas import BulkRoundRequest, ProductUpdate
+from services.manager_catalog_service import ManagerCatalogService
+
+
+router = APIRouter(prefix="/api/manager", tags=["manager"])
+
+
+@router.get("/products/list", operation_id="get_manager_products")
+async def list_products_for_manager(
+    page: int = Query(1, ge=1),
+    limit: int = Query(40, ge=1, le=100),
+    search: Optional[str] = Query(None),
+    is_published: Optional[bool] = Query(None),
+    area_min: Optional[int] = Query(None),
+    area_max: Optional[int] = Query(None),
+    is_inverter: Optional[bool] = Query(None),
+    sort: str = Query("newest"),
+    session: AsyncSession = Depends(get_session),
+    _user: str = Depends(get_current_username),
+):
+    """
+    Paginated product list for manager UI.
+    Unlike the public catalog, this can show unpublished products.
+    """
+    return await ManagerCatalogService.list_products(
+        session=session,
+        page=page,
+        limit=limit,
+        search=search,
+        is_published=is_published,
+        area_min=area_min,
+        area_max=area_max,
+        is_inverter=is_inverter,
+        sort=sort,
+    )
+
+
+@router.get("/customers", operation_id="get_manager_customers")
+async def list_customers_for_manager(
+    page: int = Query(1, ge=1),
+    limit: int = Query(20, ge=1, le=100),
+    search: Optional[str] = Query(None),
+    customer_type: Optional[str] = Query(None, alias="type"),
+    only_with_orders: bool = Query(True),
+    session: AsyncSession = Depends(get_session),
+    _user: str = Depends(get_current_username),
+):
+    """
+    Paginated customer list for manager UI.
+    Includes order count per customer.
+    """
+    return await ManagerCatalogService.list_customers(
+        session=session,
+        page=page,
+        limit=limit,
+        search=search,
+        customer_type=customer_type,
+        only_with_orders=only_with_orders,
+    )
+
+
+@router.patch("/products/{product_id}", operation_id="update_product")
+async def update_product(
+    product_id: int,
+    data: ProductUpdate,
+    session: AsyncSession = Depends(get_session),
+    _user: str = Depends(get_current_username),
+):
+    """
+    Update individual product fields.
+    """
+    result = await ManagerCatalogService.update_product(
+        session=session,
+        product_id=product_id,
+        data=data,
+    )
+
+    if not result:
+        raise HTTPException(status_code=404, detail="Product not found")
+
+    return result
+
+
+@router.post("/products/bulk-round-price", operation_id="bulk_round_price")
+async def bulk_round_price(
+    request: BulkRoundRequest,
+    session: AsyncSession = Depends(get_session),
+    _user: str = Depends(get_current_username),
+):
+    """
+    Round prices down to the nearest multiple of 50.
+    """
+    return await ManagerCatalogService.bulk_round_prices(session=session, request=request)
+
+
+@router.get("/tags/all", operation_id="get_all_tags")
+async def get_all_tags(
+    session: AsyncSession = Depends(get_session),
+    _user: str = Depends(get_current_username),
+):
+    """
+    Return all tags grouped by TagGroup for the product editor.
+    """
+    return await ManagerCatalogService.get_all_tags(session)
